Rename hive to hive chain for clarity
