+++ src/repositories/hive-chain/hive-chain.repository.spec.ts
@@ -1,5 +1,5 @@
 
-import { HiveRepository } from './hive.repository';
+import { HiveRepository } from './hive-chain.repository';
 import { ExtendedAccount, PrivateKey } from '@hiveio/dhive';
 import * as crypto from "crypto";
