+++ src/services/uploader/uploading.controller.ts
@@ -24,7 +24,7 @@ import { UploadThumbnailUploadDto } from './dto/upload-thumbnail.dto';
 import { UpdateUploadDto } from './dto/update-upload.dto';
 import { StartEncodeDto } from './dto/start-encode.dto';
 import { UploadingService } from './uploading.service';
-import { HiveRepository } from '../../repositories/hive/hive.repository';
+import { HiveRepository } from '../../repositories/hive-chain/hive-chain.repository';
 import { Upload } from './uploading.types';
 import { parseAndValidateRequest } from '../auth/auth.utils';
